moved translate function to index.js
