+++ index.js
@@ -87,3 +87,27 @@ module.exports.getWordInformation = async function (word) {
     //return result
     return returnForm;
 };
+
+
+module.exports.translate = async function (sentence, callback) {
+    //init result object
+    var res = {}
+    
+    //prepare sentence as an array
+    sentence = sentence.split(',').join('').split('.').join('').split(' ')
+    
+    //get sentence word length
+    var length = sentence.length
+    
+    //loop through words
+    sentence.forEach(async (w, key) => {
+        //get word info
+        var wordInfo = await getWordInformation(w) ? await getWordInformation(w) : w
+    
+        //put word info into result array
+        res[key] = wordInfo
+    
+        //print object when all translations fetched
+        if(Object.keys(res).length == length) await callback(res)
+    })
+}
